+++ src/stelnet/board/exploration/button/EnumButton.java
@@ -1,46 +1,47 @@
-package stelnet.board.exploration;
+package stelnet.board.exploration.button;
 
-import com.fs.starfarer.api.campaign.FactionAPI;
 import com.fs.starfarer.api.campaign.comm.IntelInfoPlugin;
 import com.fs.starfarer.api.ui.ButtonAPI.UICheckboxSize;
-import com.fs.starfarer.api.ui.Fonts;
 import com.fs.starfarer.api.ui.TooltipMakerAPI;
-import com.fs.starfarer.api.util.Misc;
 import lombok.RequiredArgsConstructor;
+import stelnet.board.exploration.ActionFilterIntel;
+import stelnet.board.exploration.ExplorationBoard;
+import stelnet.board.exploration.ExplorationL10n;
+import stelnet.board.exploration.UpdateMemoryFlag;
+import stelnet.util.L10n;
 import stelnet.util.MemoryHelper;
 import uilib2.Drawable;
 import uilib2.UiConstants;
 import uilib2.button.Button;
 import uilib2.button.ButtonBuilder;
-import uilib2.button.CheckboxCustom;
+import uilib2.button.CheckboxBasic;
 import uilib2.intel.ActionUpdateForItem;
 import uilib2.intel.ActionUpdateIntelList;
 import uilib2.intel.IntelCallbackBuilder;
 
 @RequiredArgsConstructor
-public class FactionButton extends ExplorationButton {
+public class EnumButton extends ExplorationButton {
 
-    private final FactionAPI faction;
+    private final ExplorationL10n buttonType;
     private final IntelInfoPlugin intel;
     private final float width;
     private final boolean withShift;
 
     @Override
     public void draw(TooltipMakerAPI tooltip) {
-        IdAware key = new PromotedFaction(faction);
         String memoryKeyChecked = MemoryHelper.key(
             ExplorationBoard.MEMORY_PREFIX,
-            key,
+            buttonType,
             ExplorationBoard.MEMORY_SUFFIX_CHECKED
         );
-        String memoryKeyEnabled = getMemoryKeyEnabled(key);
+        String memoryKeyEnabled = getMemoryKeyEnabled(buttonType);
         boolean isChecked = MemoryHelper.getBoolean(memoryKeyChecked, true);
         boolean isEnabled = MemoryHelper.getBoolean(memoryKeyEnabled, true);
         Drawable button = new ButtonBuilder(getButton(memoryKeyChecked, isChecked))
             .setChecked(isChecked)
             .setEnabled(isEnabled)
+            .setShortcut(buttonType.getShortcut(), false)
             .build();
-
         if (withShift) {
             button = new ShiftedButton(button, width);
         }
@@ -48,20 +49,18 @@ public class FactionButton extends ExplorationButton {
     }
 
     private Button getButton(String memoryKeyChecked, boolean isChecked) {
-        return new CheckboxCustom(
+        return new CheckboxBasic(
             (width - UiConstants.BUTTON_PADDING) / 2,
             UiConstants.BUTTON_HEIGHT,
-            Misc.ucFirst(faction.getDisplayName()),
+            L10n.get(buttonType),
             new IntelCallbackBuilder()
                 .addConfirmAction(new UpdateMemoryFlag(memoryKeyChecked, !isChecked))
                 .addConfirmAction(new ActionFilterIntel())
                 .addConfirmAction(new ActionUpdateIntelList())
                 .addConfirmAction(new ActionUpdateForItem(intel))
                 .build(),
-            Fonts.DEFAULT_SMALL,
-            faction.getBaseUIColor(),
             UICheckboxSize.SMALL,
-            UiConstants.BUTTON_PADDING
+            0
         );
     }
 }
